populated with one puppy and solve minor bug in database_setup.py

+++ database_setup.py
@@ -1,5 +1,5 @@
 import sys
-from sqlalchemy import Column, ForeignKey, Integer, String
+from sqlalchemy import Column, ForeignKey, Integer, String, Float
 from sqlalchemy.ext.declarative import declarative_base;
 from sqlalchemy.orm import relationship
 from sqlalchemy import create_engine
@@ -54,7 +54,7 @@ class Puppy(Base):
 		)
 
 	weight = Column(
-			Double,
+			Float,
 			nullable = False
 		)
